refactor: AsyncRestPollingLoopImpl template location to reduce compile time
