+++ google/cloud/google_cloud_cpp_rest_protobuf_internal.bzl
@@ -18,7 +18,10 @@
 
 google_cloud_cpp_rest_protobuf_internal_hdrs = [
     "internal/async_rest_long_running_operation.h",
+    "internal/async_rest_long_running_operation_custom.h",
     "internal/async_rest_polling_loop.h",
+    "internal/async_rest_polling_loop_custom.h",
+    "internal/async_rest_polling_loop_impl.h",
     "internal/async_rest_retry_loop.h",
     "internal/rest_background_threads_impl.h",
     "internal/rest_completion_queue_impl.h",
