+++ google/cloud/google_cloud_cpp_rest_protobuf_internal.cmake
@@ -18,8 +18,11 @@
 add_library(
     google_cloud_cpp_rest_protobuf_internal # cmake-format: sort
     internal/async_rest_long_running_operation.h
+    internal/async_rest_long_running_operation_custom.h
     internal/async_rest_polling_loop.cc
     internal/async_rest_polling_loop.h
+    internal/async_rest_polling_loop_custom.h
+    internal/async_rest_polling_loop_impl.h
     internal/async_rest_retry_loop.h
     internal/rest_background_threads_impl.cc
     internal/rest_background_threads_impl.h
@@ -114,7 +117,9 @@ if (BUILD_TESTING)
     # List the unit tests, then setup the targets and dependencies.
     set(google_cloud_cpp_rest_protobuf_internal_unit_tests
         # cmake-format: sort
+        internal/async_rest_long_running_operation_custom_test.cc
         internal/async_rest_long_running_operation_test.cc
+        internal/async_rest_polling_loop_custom_test.cc
         internal/async_rest_polling_loop_test.cc
         internal/async_rest_retry_loop_test.cc
         internal/rest_background_threads_impl_test.cc
