+++ google/cloud/internal/async_rest_long_running_operation.h
@@ -172,84 +172,6 @@ future<StatusOr<ReturnType>> AsyncRestLongRunningOperation(
       });
 }
 
-/*
- * AsyncRestLongRunningOperation for services that do not conform to AIP-151.
- */
-template <typename ReturnType, typename OperationType,
-          typename GetOperationRequestType, typename CancelOperationRequestType,
-          typename RequestType, typename StartFunctor, typename RetryPolicyType,
-          typename CompletionQueue>
-future<StatusOr<ReturnType>> AsyncRestLongRunningOperation(
-    CompletionQueue cq, internal::ImmutableOptions options,
-    RequestType&& request, StartFunctor&& start,
-    AsyncRestPollLongRunningOperation<OperationType, GetOperationRequestType>
-        poll,
-    AsyncRestCancelLongRunningOperation<CancelOperationRequestType> cancel,
-    LongRunningOperationValueExtractor<ReturnType, OperationType>
-        value_extractor,
-    std::unique_ptr<RetryPolicyType> retry_policy,
-    std::unique_ptr<BackoffPolicy> backoff_policy, Idempotency idempotent,
-    std::unique_ptr<PollingPolicy> polling_policy, char const* location,
-    std::function<bool(OperationType const&)> is_operation_done,
-    std::function<void(std::string const&, GetOperationRequestType&)>
-        get_request_set_operation_name,
-    std::function<void(std::string const&, CancelOperationRequestType&)>
-        cancel_request_set_operation_name) {
-  auto operation =
-      AsyncRestRetryLoop(std::move(retry_policy), std::move(backoff_policy),
-                         idempotent, cq, std::forward<StartFunctor>(start),
-                         options, std::forward<RequestType>(request), location);
-  auto loc = std::string{location};
-  return AsyncRestPollingLoop<OperationType, GetOperationRequestType,
-                              CancelOperationRequestType>(
-             std::move(cq), std::move(options), std::move(operation),
-             std::move(poll), std::move(cancel), std::move(polling_policy),
-             std::move(location), is_operation_done,
-             get_request_set_operation_name, cancel_request_set_operation_name)
-      .then([value_extractor, loc](future<StatusOr<OperationType>> g) {
-        return value_extractor(g.get(), loc);
-      });
-}
-
-// TODO(#12359) - remove once it is no longer used
-template <typename ReturnType, typename OperationType,
-          typename GetOperationRequestType, typename CancelOperationRequestType,
-          typename RequestType, typename StartFunctor, typename RetryPolicyType,
-          typename CompletionQueue>
-future<StatusOr<ReturnType>> AsyncRestLongRunningOperation(
-    CompletionQueue cq, RequestType&& request, StartFunctor&& start,
-    AsyncRestPollLongRunningOperationImplicitOptions<OperationType,
-                                                     GetOperationRequestType>
-        poll,
-    AsyncRestCancelLongRunningOperationImplicitOptions<
-        CancelOperationRequestType>
-        cancel,
-    LongRunningOperationValueExtractor<ReturnType, OperationType>
-        value_extractor,
-    std::unique_ptr<RetryPolicyType> retry_policy,
-    std::unique_ptr<BackoffPolicy> backoff_policy, Idempotency idempotent,
-    std::unique_ptr<PollingPolicy> polling_policy, char const* location,
-    std::function<bool(OperationType const&)> is_operation_done,
-    std::function<void(std::string const&, GetOperationRequestType&)>
-        get_request_set_operation_name,
-    std::function<void(std::string const&, CancelOperationRequestType&)>
-        cancel_request_set_operation_name) {
-  auto operation =
-      AsyncRestRetryLoop(std::move(retry_policy), std::move(backoff_policy),
-                         idempotent, cq, std::forward<StartFunctor>(start),
-                         std::forward<RequestType>(request), location);
-  auto loc = std::string{location};
-  return AsyncRestPollingLoop<OperationType, GetOperationRequestType,
-                              CancelOperationRequestType>(
-             std::move(cq), std::move(operation), std::move(poll),
-             std::move(cancel), std::move(polling_policy), std::move(location),
-             is_operation_done, get_request_set_operation_name,
-             cancel_request_set_operation_name)
-      .then([value_extractor, loc](future<StatusOr<OperationType>> g) {
-        return value_extractor(g.get(), loc);
-      });
-}
-
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace rest_internal
 }  // namespace cloud
