+++ google/cloud/internal/async_rest_long_running_operation_test.cc
@@ -541,223 +541,6 @@ TEST(AsyncLongRunningTest, RequestPollThenCancelImplicitOptions) {
   EXPECT_THAT(actual, StatusIs(StatusCode::kCancelled));
 }
 
-class BespokeOperationType {
- public:
-  bool is_done() const { return is_done_; }
-  void set_is_done(bool is_done) { is_done_ = is_done; }
-  std::string const& name() const { return name_; }
-  void set_name(std::string name) { name_ = std::move(name); }
-  std::string* mutable_name() { return &name_; }
-  bool operator==(BespokeOperationType const& other) const {
-    return is_done_ == other.is_done_ && name_ == other.name_;
-  }
-
- private:
-  bool is_done_;
-  std::string name_;
-};
-
-class BespokeGetOperationRequestType {
- public:
-  std::string const& name() const { return name_; }
-  void set_name(std::string name) { name_ = std::move(name); }
-
- private:
-  std::string name_;
-};
-
-class BespokeCancelOperationRequestType {
- public:
-  std::string const& name() const { return name_; }
-  void set_name(std::string name) { name_ = std::move(name); }
-
- private:
-  std::string name_;
-};
-
-class MockBespokeOperationStub {
- public:
-  MOCK_METHOD(future<StatusOr<BespokeOperationType>>, AsyncCreateResponse,
-              (CompletionQueue & cq, std::unique_ptr<RestContext> context,
-               Options const&, Request const&),
-              ());
-
-  MOCK_METHOD(future<StatusOr<BespokeOperationType>>, AsyncGetOperation,
-              (CompletionQueue&, std::unique_ptr<RestContext>, Options const&,
-               BespokeGetOperationRequestType const&),
-              ());
-
-  MOCK_METHOD(future<Status>, AsyncCancelOperation,
-              (CompletionQueue&, std::unique_ptr<RestContext>, Options const&,
-               BespokeCancelOperationRequestType const&),
-              ());
-};
-
-TEST(AsyncLongRunningTest,
-     RequestPollThenSuccessResponseWithBespokeOperationTypes) {
-  Response expected;
-  expected.set_seconds(123456);
-  BespokeOperationType starting_op;
-  starting_op.set_name("test-op-name");
-  starting_op.set_is_done(false);
-  BespokeOperationType done_op = starting_op;
-  done_op.set_is_done(true);
-
-  auto mock_cq = std::make_shared<MockCompletionQueueImpl>();
-  EXPECT_CALL(*mock_cq, MakeRelativeTimer)
-      .WillOnce([](std::chrono::nanoseconds) {
-        return make_ready_future(
-            make_status_or(std::chrono::system_clock::now()));
-      });
-  CompletionQueue cq(mock_cq);
-
-  auto mock = std::make_shared<MockBespokeOperationStub>();
-  EXPECT_CALL(*mock, AsyncCreateResponse)
-      .WillOnce([&](CompletionQueue&, std::unique_ptr<RestContext>,
-                    Options const& options, Request const&) {
-        EXPECT_EQ(options.get<StringOption>(), CurrentTestName());
-        return make_ready_future(make_status_or(starting_op));
-      });
-  EXPECT_CALL(*mock, AsyncGetOperation)
-      .WillOnce([&](CompletionQueue&, std::unique_ptr<RestContext>,
-                    Options const& options,
-                    BespokeGetOperationRequestType const&) {
-        EXPECT_EQ(options.get<StringOption>(), CurrentTestName());
-        return make_ready_future(make_status_or(done_op));
-      });
-  auto policy = std::make_unique<MockPollingPolicy>();
-  EXPECT_CALL(*policy, clone()).Times(0);
-  EXPECT_CALL(*policy, OnFailure).Times(0);
-  EXPECT_CALL(*policy, WaitPeriod)
-      .WillRepeatedly(Return(std::chrono::milliseconds(1)));
-  Request request;
-  request.set_seconds(123456);
-  request.set_nanos(456789);
-
-  auto options = internal::MakeImmutableOptions(
-      Options{}.set<StringOption>(CurrentTestName()));
-  auto actual =
-      AsyncRestLongRunningOperation<Response, BespokeOperationType,
-                                    BespokeGetOperationRequestType,
-                                    BespokeCancelOperationRequestType>(
-          cq, options, std::move(request),
-          [mock](CompletionQueue& cq, std::unique_ptr<RestContext> context,
-                 Options const& options, Request const& request) {
-            return mock->AsyncCreateResponse(cq, std::move(context), options,
-                                             request);
-          },
-          [mock](CompletionQueue& cq, std::unique_ptr<RestContext> context,
-                 Options const& options,
-                 BespokeGetOperationRequestType const& request) {
-            return mock->AsyncGetOperation(cq, std::move(context), options,
-                                           request);
-          },
-          [mock](CompletionQueue& cq, std::unique_ptr<RestContext> context,
-                 Options const& options,
-                 BespokeCancelOperationRequestType const& request) {
-            return mock->AsyncCancelOperation(cq, std::move(context), options,
-                                              request);
-          },
-          [&](StatusOr<BespokeOperationType> const&,
-              std::string const&) -> StatusOr<Response> { return expected; },
-          TestRetryPolicy(), TestBackoffPolicy(), Idempotency::kIdempotent,
-          std::move(policy), "test-function",
-          [](BespokeOperationType const& op) { return op.is_done(); },
-          [](std::string const& s, BespokeGetOperationRequestType& op) {
-            op.set_name(s);
-          },
-          [](std::string const& s, BespokeCancelOperationRequestType& op) {
-            op.set_name(s);
-          })
-          .get();
-  internal::OptionsSpan overlay(Options{}.set<StringOption>("uh-oh"));
-  ASSERT_THAT(actual, IsOk());
-  EXPECT_THAT(*actual, IsProtoEqual(expected));
-}
-
-TEST(AsyncLongRunningTest,
-     RequestPollThenSuccessResponseWithBespokeOperationTypesImplicitOptions) {
-  Response expected;
-  expected.set_seconds(123456);
-  BespokeOperationType starting_op;
-  starting_op.set_name("test-op-name");
-  starting_op.set_is_done(false);
-  BespokeOperationType done_op = starting_op;
-  done_op.set_is_done(true);
-
-  auto mock_cq = std::make_shared<MockCompletionQueueImpl>();
-  EXPECT_CALL(*mock_cq, MakeRelativeTimer)
-      .WillOnce([](std::chrono::nanoseconds) {
-        return make_ready_future(
-            make_status_or(std::chrono::system_clock::now()));
-      });
-  CompletionQueue cq(mock_cq);
-
-  auto mock = std::make_shared<MockBespokeOperationStub>();
-  EXPECT_CALL(*mock, AsyncCreateResponse)
-      .WillOnce([&](CompletionQueue&, std::unique_ptr<RestContext>,
-                    Options const& options, Request const&) {
-        EXPECT_EQ(internal::CurrentOptions().get<StringOption>(),
-                  CurrentTestName());
-        EXPECT_EQ(options.get<StringOption>(), CurrentTestName());
-        return make_ready_future(make_status_or(starting_op));
-      });
-  EXPECT_CALL(*mock, AsyncGetOperation)
-      .WillOnce([&](CompletionQueue&, std::unique_ptr<RestContext>,
-                    Options const& options,
-                    BespokeGetOperationRequestType const&) {
-        EXPECT_EQ(internal::CurrentOptions().get<StringOption>(),
-                  CurrentTestName());
-        EXPECT_EQ(options.get<StringOption>(), CurrentTestName());
-        return make_ready_future(make_status_or(done_op));
-      });
-  auto policy = std::make_unique<MockPollingPolicy>();
-  EXPECT_CALL(*policy, clone()).Times(0);
-  EXPECT_CALL(*policy, OnFailure).Times(0);
-  EXPECT_CALL(*policy, WaitPeriod)
-      .WillRepeatedly(Return(std::chrono::milliseconds(1)));
-  Request request;
-  request.set_seconds(123456);
-  request.set_nanos(456789);
-
-  internal::OptionsSpan span(Options{}.set<StringOption>(CurrentTestName()));
-  auto actual =
-      AsyncRestLongRunningOperation<Response, BespokeOperationType,
-                                    BespokeGetOperationRequestType,
-                                    BespokeCancelOperationRequestType>(
-          cq, std::move(request),
-          [mock](CompletionQueue& cq, std::unique_ptr<RestContext> context,
-                 Request const& request) {
-            return mock->AsyncCreateResponse(
-                cq, std::move(context), internal::CurrentOptions(), request);
-          },
-          [mock](CompletionQueue& cq, std::unique_ptr<RestContext> context,
-                 BespokeGetOperationRequestType const& request) {
-            return mock->AsyncGetOperation(cq, std::move(context),
-                                           internal::CurrentOptions(), request);
-          },
-          [mock](CompletionQueue& cq, std::unique_ptr<RestContext> context,
-                 BespokeCancelOperationRequestType const& request) {
-            return mock->AsyncCancelOperation(
-                cq, std::move(context), internal::CurrentOptions(), request);
-          },
-          [&](StatusOr<BespokeOperationType> const&,
-              std::string const&) -> StatusOr<Response> { return expected; },
-          TestRetryPolicy(), TestBackoffPolicy(), Idempotency::kIdempotent,
-          std::move(policy), "test-function",
-          [](BespokeOperationType const& op) { return op.is_done(); },
-          [](std::string const& s, BespokeGetOperationRequestType& op) {
-            op.set_name(s);
-          },
-          [](std::string const& s, BespokeCancelOperationRequestType& op) {
-            op.set_name(s);
-          })
-          .get();
-  internal::OptionsSpan overlay(Options{}.set<StringOption>("uh-oh"));
-  ASSERT_THAT(actual, IsOk());
-  EXPECT_THAT(*actual, IsProtoEqual(expected));
-}
-
 }  // namespace
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace rest_internal
